+++ CCPS610-Assignment2/DatabaseContext/Tables/HrEmployee.cs
@@ -22,7 +22,7 @@ namespace CCPS610_Assignment2.DatabaseContext.Tables
         public decimal? Salary { get; set; }
         public decimal? CommissionPct { get; set; }
         public int? ManagerId { get; set; }
-        public byte? DepartmentId { get; set; }
+        public int? DepartmentId { get; set; }
 
         public virtual HrDepartment? Department { get; set; }
         public virtual HrJob Job { get; set; } = null!;
